refactor(coords): rename constants.py to coords.py
