+++ pyclashbot/bot/nav.py
@@ -2,7 +2,7 @@ import random
 import time
 from typing import Literal
 
-from pyclashbot.bot.constants import CLASH_MAIN_DEADSPACE_COORD as CLASH_MAIN_MENU_DEADSPACE_COORD
+from pyclashbot.bot.coords import CLASH_MAIN_DEADSPACE_COORD as CLASH_MAIN_MENU_DEADSPACE_COORD
 from pyclashbot.bot.state_detect import (
     check_for_trophy_reward_menu,
     check_if_in_battle,
